[IMP] pos: Show categories on mobile
Before this commit, users had to use a popup window to change
the product category. It wasn't clear for users who are used
to using a large screen.

This commit adapts the breadcrumb navigation to be the same
as on desktop. This avoids the use of this popup.

We also improve the display of products by taking the entire width
of the screen. This allows us to add some space between the products.

Task ID: 2355025

Part-of: odoo/odoo#78157
Co-authored-by: Adrien Dieudonné <adr@odoo.com>