[IMP] mail: avoid email loops when Odoo reply automatically
Purpose
=======
In this commit, we will fix the loops that occur when you send an email
to an alias (to create a ticket e.g.). In that case Odoo can reply
"Your ticket has been created" and then the auto-replier of the user
can reply to this email and the loop occurs.

Specifications
==============
To solve this issues, we add 2 system parameters
- <mail.incoming.limit.period>, 60 minutes by default
- <mail.incoming.limit.alias>, 5 by default

When an email is sent to an alias, we look on the last records created
<mail.incoming.limit.period> minutes ago. If we overcome the limit
<mail.incoming.limit.alias> the email is ignored.

Alias creation detection
========================
To detect the number of records created by a specific email address we
use the `_primary_email` attribute set on the model.

Task-2294034