[FIX] l10n_uk: Update UK tax report, taxes and fiscal positions
The tax report for the UK is outdated, in this commit it is updated
to reflect the changes made to it after Brexit. New taxes are added
to better cover the standard taxes found in the UK. Some UK demo data
is also removed in this commit as it is generally not useful. Fiscal
positions are added that cover Great Britain's trading situtation with
the rest of the world. Northern Ireland's specific taxes, fiscal
positions, chart of accounts will are added in another localisation
[l10n_xi] in a different commit.

task-2675406