http.py _serve_static
The `send_file` utility have been split in two: `send_filepath` that
takes a path-like and `send_file` that takes either a file-objet or
the content right away (as str/bytes). This new `send_filepath` is a
request from the Odoo SH team and should be the preferred way to stream
a file located on the filesystem. The original function has been
deprecated to be moved on Request, Antony prefers it this way, I
personally don't really care.

We wanted to define `Application.statics` as a lazy-property but the
magic global variable `addons_manifest` needs to be populated upon the
first query. This magic global variable is basically an addons manifest
cache, it should be defined and populated elsewhere (e.g. in
`modules/module.py`) but that's duty for another PR.

You'll notice that `_serve_static` raises `werkzeug.exceptions.NotFound`
in the case the requested file couldn't be served. You may wonder how we
handle exceptions, it is in a future commit.