http.py _serve_db, model.retrying (model.check), session
The session is now located on the database. Because the session is using
the database, the dbname can non longer be retrieved using only the
session identifier. Each request requiring the database must now provide
it. The `session_id` cookie format has been updated to also include the
database:

    <session id>.<database name>

Maybe we could save the database in another cookie? It works well for
now but I'm not too much satisfied by mixing two values into one cookie.

The ORM initialization (cursor, registry, environment) was pretty much
magic before, it is now completely deterministic. There is a new
function `request.update_env` (and its companion `update_context`) that
should be used to replace the environment attached to the current
request.

The previous implementation was using some magical properties to renew
the environment each time one modified the uid or the context. It was
difficult to reason about where and when was the environment re-created.

The `odoo.service.model.check` decorator have been refactored too, the
function is now called `retrying` which better describes its purpose. It
is no-longer a decorator so it is possible to "retry" any function.