[IMP] calendar: add discuss video url in calendar
This commit aims to integrate the calendar module with discuss video calls.
Now it is possible to choose the videocall source as either discuss or custom.
Custom source allows the user to enter any URL, while discuss source generates
an URL that will be used to create the discuss channel.
For recurring events, the same discuss channel is used for all events.
For single events, each event will create a discuss channel.
The discuss channel is only created when someone tries to enter the videocall
URL.

[FIX] calendar: title undefined when double clicking event

Currently in calendar if you double click an event in the calendar view, it
opens a dialog with the event title as undefined. This occurs, because the
title is not correctly passed in the context. This commit adds the correct
title to it, therefore fixing the error.

task-2685472