[IMP] website: add a hover effect option
Add an option in the editor to activate custom content when hovering
specific elements. Using this option adds an invisible overlay on top
of the original element that will only be shown to the visitor when
hovering the original element.

In edit mode the hovering behavior is deactivated, so both the original
and the overlay content can be properly edited. Buttons are provided
that allow the end user to quickly toggle between the original and
overlay content.

task-2666891

Co-Authored-By: Younn Olivier <yol@odoo.com>