[REF] rating: improve rating from routes
RATIONALE
Currently, rating from email (route with token) is the following

  * Customer receives an email;
  * Customer clicks one of the smiley face that redirects to a ``/rate`` route
    with the token and rate value;
  * the linked ``rating.rating`` record is updated and set consumed;
  * Customer may change its face and add a feedback;
  * in that case a message is posted with the rating face in it;

Issues

  * first rating update does not notify anyone, no message is posted;
  * if Customer does not give a feedback, process is stuck at that point;
  * the rating is not really linked to the posted message (see ``message_id``
    field of ``rating`` record);

In this commit we change that flow. Now the fow is

SPECIFICATIONS

Update ``rating_apply`` to either post a rating or update the message with
the new value and/or feedback. Always link the message and the rating to
be sure DB data are coherent.

CODE CLEANUP

Split behavior that is generic but actually used only in project. Move it
directly in project. It is used to update kanban state of tasks based on
rating. As it is used only once and as generic approach is hardocoded based
on fields names, this is not really loosing features.