[IMP] hr_holidays: automatic time-off recovery after public holiday edit
If a public holiday is created, modified or deleted, all declared (non-refused) leaves related to that period will have
their duration re-evaluated and individual remaining time offs will be computed accordingly.
If the duration of the leave is evaluated at 0 after the creation of the public holiday, the leave will
be set as cancelled.

task-2636215
[Time Off]Grant back time off when public day off is created