[FIX] core: use EXISTS in domain conditions like (m2m, '=', False)
Consider the domain [('tag_ids', '=', False)] with a many2many field.
It is currently translated in SQL as

    "model".id NOT IN (
        SELECT "model_id"
        FROM "model_tag_rel"
        WHERE "model_id" IS NOT NULL
    )

PostgreSQL does not handle well this "NOT IN" expression when the
relation table is large, and uses some very inefficient query plan in
that case.  This commit changes the above expression to a "NOT EXISTS"
expression (as below), which is executed with a much more efficient
query plan, even with large tables.

    NOT EXISTS (
        SELECT 1
        FROM "model_tag_rel"
        WHERE "model_tag_rel"."model_id" = "model".id
    )

This change was motivated by a use-case with 2.7M account moves, 11M
account move lines and a relation table `account_move_line_account_tax_rel`
with 2.3M lines.  The query was timing out (15 minutes), and it now
takes 15 seconds.