[IMP] *: Apply changes following the HTTPocalypse
Public API changes
------------------

*You care about it*: you always do.

* `request.uid = x` => `request.update_env(user=x)`.
* `request.context = x` => `request.update_env(context=x)`.
* `request.context = dict(request.context, x=y)`
  => `request.update_context(x=y)`.
* `request.cr = None` => `request.cr.close()`.
* `http.mono_db()` => `request.db`.
* `http.send_file` => `request.send_filepath`/`request.send_file`.
* `http.dispatch_rpc` => `service.dispatch_rpc`.
* `service.model.check` decorator => `service.model.retrying` wrapper.
* `request.params` => set very late, in case of `AttributeError` in any
  `ir.http` override please use `request.get_http_params` (query-string
  + form) or `request.httprequest.args` (query-string) instead.
* `request.endpoint` => removed, call `_match()` instead. In case you
  *absolutely* need it back, please reach us in the github issues.
* `request.routing_iteration` => removed, not necessary anymore.
* `@route` => the `type` argument for new endpoints is now mandatory.

New APIs
--------

*You care about it*: you always do.

The session is now json-serialized and stored in the database thanks to
the new `ir.session` model. The session of the user is only override
when its current session at the end of the request is different from the
one stored.

Thanks to `request.future_response` it is possible to set headers and
cookies on the response object *before* the response object is
initialized. Please note that headers/cookies saved on the future
response will NOT be injected in case of error.

ir.http changes
---------------

*You care about it*: you were overriding `ir.http`.

The call-stack was as follow:

    http/Application@_dispatch
    -> ir.http@_dispatch
       -> ir.http@_match
       -> ir.http@_authenticate
       -> ir.http@_postprocess_args
       -> http/Request@dispatch
          ...
          -> @route decorated controller method

It is now as follow when serving backend routes:

    http/Request/_serve_db
    -> http/Request/_serve_ir_http
       -> ir.http@_match
       -> ir.http@_authenticate
       -> ir.http@_pre_dispatch
          -> ir.http@_frontend_pre_dispatch  # @route(website=True)
       -> http/Request/_http_dispatch
          -> ir.http@_dispatch
             -> @route decorated controller method

It is now as follow when serving frontend pages:

    http/Request/_serve_db
    -> http/Request/_serve_ir_http
       -> ir.http@_match  # raises 404 not found
       -> ir.http@_serve_fallback
          -> ir.http@_auth_method_public
          -> ir.http@_frontend_pre_dispatch
          -> ir.http@_serve_page

ir.http _match()
----------------

*You care about it*: you need to modify the request path before you
match a backend route and you cannot redirect (3xx response) the user on
that URL.

See `_match` override in `http_routing` for details about internal
rerouting.

ir.http _auth_method_x
----------------------

*You care about it*: you need to grant the access to *some* unregistered
users but cannot ask for a user login/pwd/2fa nor you can set the route
public nor it is possible for the users to access the backend via RPC.

See `_auth_method_calendar` in `calendar`.

ir.http _pre_dispatch
---------------------

*You care about it*: you need to save some options from the query-string
in the session, in the context or in the response cookies before serving
a **backend** endpoint.

See `_pre_dispatch` override in `web`.

Important note: this method is only called in the context of a backend
request, a request that did match a route.

ir.http _frontend_pre_dispatch
------------------------------

*You care about it*: you need to save some options from the query-string
in the session, in the context or in the response cookies before serving
a backend `@route(website=True)` endpoint or a **frontend** page.

See `_frontend_pre_dispatch` override in `website`.

ir.http _serve_fallback
-----------------------

*You care about it*: you need to deliver a new kind of content, content
that cannot be served nor via the static files, nor via a backend
endpoint nor via portal/website but you still need Odoo.

This is discouraged.

In case you can affort *not to use Odoo* or to use Odoo via RPC, start
*another* web application and configure your web server (apache, nginx)
to route some requests there.

In case you don't, be aware that Odoo needs to match a backend route in
order to correctly setup the ORM. Overriding `_serve_fallback` puts you
in a situation where the request **IS NOT** authenticated and **NO**
user is set in the environment. You'll have to come with your own
authentication scheme and to set a user on the environment attached to
the request before using the ORM.

See `_serve_fallback` in `http_routing`.

ir.http _http_handle_error
---------------------

*You care about it*: you need a custom behavior in case of errors.

Make sure there is a user configured in the request environment
otherwise if you need to use the ORM:

    if not request.uid:
        cls._auth_method_public()  # grant the public user

See `_http_handle_error` in `http_routing`.