[IMP] digest: fix tip description
Before this Commit,
While sending the digest emails, The tip description has improper
because of sanitizing.

After this Commit,
We have fixed the tip description field value. We do the html_sanitize
after rendering the QWEB template. So, we get the proper values of conditions
and aliases. Replacing jinja with qweb template, should use
<t t-out="variable"/> to print the dynamic value so, remove the inline_template
expression `{{` and `}}` and put the <t t-out="variable"/>.

Task-2704083