[REF] point_of_sale: search products not loaded yet
In the current approach, a PoS user can open a new session
loading only the most common products beforehand. The user
may choose to load the remaining products in the background
after the session has started. If not, the user can access
the remaining not loaded products by typing the desired
product name in the search bar and then pressing the database
icon button.

The current approach is not straight forward. To fix this,
in the current commit we drop the database icon and we
extend the functionality of the search bar so that when the
user presses "Enter" the database is querried for more products.

task-2669747