[REF] point_of_sale: drop the "Set Customer" button
Most of the time, the user wants to select a customer, not to perform
other actions. By dropping the "Set Customer" button the user can select
a customer directly by clicking on them. As a result this action is
completed in one click instead of two. This follows the "make the common
case fast" design principle.

task-2669747