[IMP] web_editor: only trigger a visibility update on the active snippet
A snippet_option_visibility_update can be triggered to toggle the visibility of
an invisible element. Currently the event is handled by both the editors and the
SnippetsMenu which means it will affect all parent snippets as well.

By only handling the event in the SnippetsMenu this issue can be solved. The
listener in the SnippetsMenu instead checks the target of the event to find the
affected snippet.

In case the snippet differs from the target of the editor (or even if it isn't
triggered by an editor) where the event is triggered, a $snippet argument can
be passed along with the event.