[FIX] mass_mailing: add active equals to false for a/b testing cron
Before this commit the A/B testing cron was active by default.
It doesn't matter if a/b testing is true or false in settings.
So, with this commit we set active equals to false by default
in a/b testing cron. I some one enables in settings it will
become active.

TaskId-2713198