[IMP] point_of_sale: move and rename "EDIT" button
In the ClientListScreen, we move the "EDIT" button to the right,
we rename it to "DETAILS" button and we make it always visible.
The reason for this is that the main action on this screen is
to select a client. To simplify the interface, this action is
performed in one click on the client line, so the "DETAILS"
button needs to be always visible.

Finally, we delete some unused variables and functions.

task-2669747