[IMP] point_of_sale: use short address in the client screen
There is no need for the user to see the whole address of a client
as this field is lengthy and in most use cases is not needed.
For this reason we shorten the address column so that it displays
the postal code, the city and the state of the client (only when a
state exists). We also display the company of the client just below
their name, if the client is part of a company.

task-2669747