[FIX] mass_mailing{_sms}: remove action_select_as_winner button
Before this commit there was two buttons
1) Send this version to remaining recipients
2) Send this as winner
which were visible in view when ab_testing_winner_selection
is manual. Both of them means same that to send the current
version as winner. In both button we werre calling same method
'action_select_as_winner' at the end and both are doing the
same thing functionally and techniqually. So, with this commit
we remove the button 'Send this as winner' from view.

TaskId-2713198