[IMP] sale_timesheet: add condition to click SOL name in project update
Before this commit, all SOLs shown in the project update are clickable
if they contain a service product to view the timesheets related but
some SOLs could have no timesheets related.

This commit checks if the SOL has at least a timesheet to allow the user
to click on the SOL to see the related timesheets.

task-2710808