[IMP] point_of_sale: drop "Load Customers" button
In this commit, we extend the functionality of the
"Search Customers" search bar so that each time the
user presses "Enter" the database is queried for
customers not loaded yet. This functionality, which
was previously provided by the "Load Customers" button,
is now more straight-forward and less perplexing for
the user.

Furthermore, the search bar is moved towards the right
to better match the empty space created by the dropped
"Load Customers" button.

task-2669747