[IMP] website: add new content systray item
WIP - the implementation will probably change.
As the goal of the task is to remove the website navbar, the menus have
to be reimplemented in the backend.

Here, the new_content systray item should have reorganizable components
that will allow to install a module when it is not already there. When
the module is already installed, it should extend the component to
provide with its own onClick="createNewContent" prop.

task-2687506