[IMP] sms,test_mail_full: sms composer promenade
Before this Commit,
There is an extra space at the top of the SMS composer.
The number of characters is disappearing on the focus out after the
message box. Country selection is not available in the SMS composer.

After this Commit,
we have improved the UX of the SMS composer.

0/ Remove the extra space at the top of the SMS composer
1/ Number of characters will show consistently. (Number of characters are
disappearing on the focus out after the message box. so, the user will
lose the message count information)
2/ Add the new country field and put the county code before the
phone/mobile number automatically.

Task-2710478