[IMP] portal: add portal wizard style to ensure tree width optimality
A short scss file is added as backend assets in order to make the elements
of the user_ids tree view as close to each other as possible. This prevents
in most cases the loss of buttons going off the list's width view and
generating a scroll bar, a situation to avoid since it does not show the
user the (dyamically appeared) action buttons.

Task-2613780