[IMP] point_of_sale: create new "Balance" column
Before this commit the fields "loyalty_points"
from pos_loyalty module and "total_due" from
pos_settle_due module were displayed in two
separate columns in the client list screen.
If one of these fields were 0, the quantity
zero was displayed. This was taking up too
much space without displaying any useful info.

For this reason, we merge the two fields
(loyalty points, amount due) in a new column
named "Balance". The data is displayed only
when it is non-zero and when the corresponding
addons are installed.

task-2669747