[IMP] point_of_sale: "save" automatically selects the client
In the clientListScreen, when we edit the
details of the client and then press "Save"
the client is now automatically selected
and the user is taken back to the main screen.
This makes the common case faster since, most
of the times, when the user edits a client they
also want to select them.

task-2669747