[IMP] project_sale_expense: add reinvoice expense in profitability section
Before this commit, only the cost of an expense is taken in account in
the profitability section of the right side panel of the project update.
However, an expense can be reinvoiced to the customer and so the expense
should be, in this case, in the revenues section too but the price will
be the sell price and not the one in the expense.

This commit adds the reinvoice expense in the profitability section in
the right panel of the project update.

task-2710808