[IMP] project_mrp: add MO in project profitability
Before this commit, the MO related to the project is not taken into
account, so the MO costs is not added.

This commit adds the MO related to the project in the cost section of
the project profitability shown in the right side panel of the project
update view.

task-2710808